requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
joblib==1.3.2
mangum==0.17.0
multidict==6.0.5
numpy==1.25.2
openai==1.12.0
pluggy==1.3.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
sentry-sdk==1.41.0
sniffio==1.3.0
starlette==0.27.0
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.23.2
yarl==1.9.4
